Replace macro `unwrap!()` with `const fn` (#179)
* Replace macro `unwrap!()` with `const fn`

The library compiles quite a lot faster this way.

* Pin `regex`

* Move `fuzz` into its own workspace

* Fix pedantic warnings

* Update to 2023c

* Update changelog

* bench: Use `xoroshiro128++` instead of `SmallRng`

The algo used in `SmallRng` not fixed, so it could be that different
inputs get benchmarked, making the result useless.

Also `SmallRng` does not like seeds with little popcounts. Since our
seeds are something like `0..100`, this is not quite optimal.

* Move benchmarks into their own workspace

This way its many dependencies won't turn up in the library's
`Cargo.toml`, so I can unpin many pinned dependencies, and still comply
with our msrv.